+++ changelog.md
@@ -3,6 +3,8 @@
 ## v0.2
 
 - DSL refactoring
+- `Should` renamed to `Spec` and a new method `should` was added to it to
+  improve the DSL.
 
 ## v0.1 (02-06-2014)